feat: add context, close icon and rename components
