+++ src/app/applications/[id]/fileUploaded.tsx
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-import { faFileCircleCheck } from "@fortawesome/free-solid-svg-icons";
+import { faClose, faFileCircleCheck } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
 type FileUploadedProps = {
@@ -10,6 +10,7 @@ type FileUploadedProps = {
 };
 
 function FileUploaded({ filename }: FileUploadedProps) {
+  function onDeleteAttachment() {}
   return (
     <div className="relative mt-5 flex items-center justify-between gap-x-1 rounded border-2 bg-white-smoke px-3 py-1.5 sm:gap-x-3">
       <div className="flex items-center gap-x-2 sm:gap-x-4">
@@ -19,6 +20,11 @@ function FileUploaded({ filename }: FileUploadedProps) {
         />
         <h3 className="text-sm text-info">{filename}</h3>
       </div>
+      <FontAwesomeIcon
+        icon={faClose}
+        className="border-1 text-sm text-info"
+        onClick={onDeleteAttachment}
+      />
     </div>
   );
 }
